first step of refactoring: changing data preparation ingestion
