+++ src/data_preparation_commands.py
@@ -1,7 +1,7 @@
-from data_preparation.clean_columns import CleanColumns
-from data_preparation.region_join import RegionJoin
-from data_preparation.affects import Affects
-from data_preparation.dataset_join import DatasetsJoin
+from data_combination.clean_columns import CleanColumns
+from data_combination.region_join import RegionJoin
+from data_combination.affects import Affects
+from data_combination.dataset_join import DatasetsJoin
 from data_preparation.missing_data import MissingData
 from data_preparation.region_cleaning import RegionCleaning
 from lib.commands_handler import Commands
