+++ src/services/region_service.py
@@ -1,0 +1,3 @@
+class RegionService:
+    def __init__():
+        pass
